refactor: rename to AccountModuleFacade
