+++ pkg/intermodule/timeline.ts
@@ -1,8 +1,8 @@
 import { Result } from '@mikuroxina/mini-fn';
-import type { Note } from '../../notes/model/note.js';
-import type { PushTimelineService } from '../../timeline/service/push.js';
+import type { Note } from '../notes/model/note.js';
+import type { PushTimelineService } from '../timeline/service/push.js';
 
-export class TimelineModule {
+export class TimelineModuleFacade {
   constructor(private readonly pushTimelineService: PushTimelineService) {}
 
   /*
